Some problem occur but fixed
I stuck and trying to revert changes before commit (to valid Path) but project collapsed, so I deleted project and pull Project again.
The problem was in EnemyMover Script. In IEnumerator FollowPath() ı wrote for loop wrong (for(int i = 0; path.Count < 0; i++) ), but it should be  ( for(int i = 0; i < path.Count; i++) )